[tools] perf tests evsel-tp-sched: Fix bitwise operator
Message-id: <c3e0eccb19e9442dc0476f52aa617e1f5bdb50be.1560250176.git.mpetlan@redhat.com>
Patchwork-id: 265101
O-Subject: [RHEL8.1 BZ1715507 227/234] perf tests evsel-tp-sched: Fix bitwise operator
Bugzilla: 1715507
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715507
https://bugzilla.redhat.com/show_bug.cgi?id=1715507

upstream
========
commit 489338a717a0dfbbd5a3fabccf172b78f0ac9015
Author: Gustavo A. R. Silva <gustavo@embeddedor.com>
Date: Tue Jan 22 17:34:39 2019 -0600

description
===========
Notice that the use of the bitwise OR operator '|' always leads to true
in this particular case, which seems a bit suspicious due to the context
in which this expression is being used.

Fix this by using bitwise AND operator '&' instead.

This bug was detected with the help of Coccinelle.

Fixes: 6a6cd11d4e57 ("perf test: Add test for the sched tracepoint format fields")
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>