[rpmspec] redhat: Package the signing key for s390
Message-id: <20190627101258.24777-4-prudo@redhat.com>
Patchwork-id: 265870
O-Subject: [RHEL8.1 PATCH v3 3/3] [rpmspec] redhat: Package the signing key for s390
Bugzilla: 1714194
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Hendrik Brueckner <brueckner@redhat.com>
RH-Acked-by: Jan Stancek <jstancek@redhat.com>

Bugzilla: 1714194
Upstream Status: RHEL only
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=22366678

To verify the kernel signature the public signing key is needed. Package
it together with the signing CA in the kernel package for s390.

Packaging the public key with the kernel is especially important for
scratch builds. For them the module signing key is used to sign the
kernel. This key is generated during and differs with every build. Thus
it must be packaged together with the kernel.

In the long term a signing tool similar to pesign can be used to ship the
keys. Once it exists this patch can be reverted.

Signed-off-by: Philipp Rudo <prudo@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>