[redhat] kabi: Adding symbol submit_bio (block/blk-core.c)
Message-id: <20190703120021.6957-198-ckalina@redhat.com>
Patchwork-id: 266318
O-Subject: [RHEL8.1 KABI 216/236] kabi: Adding symbol submit_bio (block/blk-core.c)
Bugzilla: 1677170
RH-Acked-by: Ming Lei <ming.lei@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

Bugzilla: 1677170
https://bugzilla.redhat.com/show_bug.cgi?id=1677170

Description:
============
pdfs_ko - The Primesoft Distributed File System for Hadoop (P-DFS) uses
this in process of read/write to submit block I/O as same as a lot of
processing in the kernel code.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

extern blk_qc_t submit_bio ( struct bio * )

Cc: Jeff Moyer <jmoyer@redhat.com>
Cc: Ming Lei <minlei@redhat.com>
Cc: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>