[redhat] kabi: Adding symbol dst_release (net/core/dst.c)
Message-id: <20190709191750.10507-32-ckalina@redhat.com>
Patchwork-id: 266473
O-Subject: [RHEL8.1 KABI 268/236] kabi: Adding symbol dst_release (net/core/dst.c)
Bugzilla: 1684393
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>

Bugzilla: 1684393
https://bugzilla.redhat.com/show_bug.cgi?id=1684393

Description:
============
cf_ko - This function is is used to free the sk_dst_cache member of the
sock structure int the network interface of CF of PRIMECLUSTER.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

void dst_release ( struct dst_entry * )

Cc: Jiri Benc <jbenc@redhat.com>
Cc: Neil Horman <nhorman@redhat.com>
Cc: Rashid Khan <rkhan@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>