RHELPatchRules: modify the official policy
This catches up RHELPatchRules to the actual process that has been observed
for a while: public patches for publicly known security issues are being
sent to rhkernel-list, not rhkernel-team-list. This practice is not
conformant to RHELPatchRules.

As it's been the current practice for a long time, codifying this pracitce
as the official policy is unlikely to cause issues.

The advantage is that unlike with rhkernel-team-list, the patches submitted
to rhkl go through submaintainers (where applicable) and are subjected to
more targeted subsystem testing.

Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>