[x86] kvm: taint kernel for tech-preview when using nested virtualization
Message-id: <20190822184037.27411-2-pbonzini@redhat.com>
Patchwork-id: 269505
O-Subject: [RHEL8.1 virt PATCH v3 1/2] KVM: taint kernel for tech-preview when using nested virtualization
Bugzilla: 1739739
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>

Upstream status: RHEL-only

For Intel, tainting the kernel happens on VMPTRLD so that simply
loading the KVM modules does not trigger it.  For AMD, setting
the module parameter taints the kernel since it is not the default.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>