[tools] selftests: Move test output to diagnostic lines
Message-id: <20190823233803.25327-7-drosario@redhat.com>
Patchwork-id: 269629
O-Subject: [RHEL-8.1 BZ1721461 PATCH 6/8] selftests: Move test output to diagnostic lines
Bugzilla: 1721461
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1721461

DESCRIPTION:
------------
This changes the selftest output so that each test's output is prefixed
with "# " as a TAP "diagnostic line".

This creates a bit of a kernel-specific TAP dialect where the diagnostics
precede the results. The TAP spec isn't entirely clear about this, though,
so I think it's the correct solution so as to keep interactive runs making
sense. If the output _followed_ the result line in the spec-suggested
YAML form, each test would dump all of its output at once instead of as
it went, making debugging harder.

This does, however, solve the recursive TAP output problem, as sub-tests
will simply be prefixed by "# ". Parsing sub-tests becomes a simple
problem of just removing the first two characters of a given top-level
test's diagnostic output, and parsing the results.

Note that the shell construct needed to both get an exit code from
the first command in a pipe and still filter the pipe (to add the "# "
prefix) uses a POSIX solution rather than the bash "pipefail" option
which is not supported by dash.

Since some test environments may have a very minimal set of utilities
available, the new prefixing code will fall back to doing line-at-a-time
prefixing if perl and/or stdbuf are not available.

Signed-off-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Shuah Khan <skhan@linuxfoundation.org>

UPSTREAM STATUS:
----------------
commit 5c069b6dedef1fab5420ca8658ed7f9ee4d26007

https://git.kernel.org/pub/scm/linux/kernel/git/powerpc/linux.git/commit/?id=5c069b6dedef1fab5420ca8658ed7f9ee4d26007

Signed-off-by: Phillip Lougher <plougher@redhat.com>