[tools] selftests/powerpc: Do not fail with reschedule
Message-id: <20190828152859.12632-2-drosario@redhat.com>
Patchwork-id: 269955
O-Subject: [RHEL-8.1 BZ1739770 PATCH 1/2] selftests/powerpc: Do not fail with reschedule
Bugzilla: 1739770
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1739770

DESCRIPTION:
------------
There are cases where the test is not expecting to have the transaction
aborted, but, the test process might have been rescheduled, either in the
OS level or by KVM (if it is running on a KVM guest machine). The process
reschedule will cause a treclaim/recheckpoint which will cause the
transaction to doom, aborting the transaction as soon as the process is
rescheduled back to the CPU. This might cause the test to fail, but this is
not a failure in essence.

If that is the case, TEXASR[FC] is indicated with either
TM_CAUSE_RESCHEDULE or TM_CAUSE_KVM_RESCHEDULE for KVM interruptions.

In this scenario, ignore these two failures and avoid the whole test to
return failure.

Signed-off-by: Breno Leitao <leitao@debian.org>
Reviewed-by: Gustavo Romero <gromero@linux.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

UPSTREAM STATUS:
----------------
commit 44d947eff19d64384efc06069509db7a0a1103b0

https://git.kernel.org/pub/scm/linux/kernel/git/powerpc/linux.git/commit/?id=44d947eff19d64384efc06069509db7a0a1103b0

Signed-off-by: Phillip Lougher <plougher@redhat.com>