[tools] kvm: selftests: Add the sync_regs test for s390x
Message-id: <20190910154757.15199-9-thuth@redhat.com>
Patchwork-id: 271197
O-Subject: [RHEL8.2 virt BZ 1730412 PATCH 08/22] KVM: selftests: Add the sync_regs test for s390x
Bugzilla: 1730412
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: David Hildenbrand <david@redhat.com>
RH-Acked-by: Peter Xu <peterx@redhat.com>

The test is an adaption of the same test for x86. Note that there
are some differences in the way how s390x deals with the kvm_valid_regs
in struct kvm_run, so some of the tests had to be removed. Also this
test is not using the ucall() interface on s390x yet (which would need
some work to be usable on s390x), so it simply drops out of the VM with
a diag 0x501 breakpoint instead.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Message-Id: <20190523164309.13345-8-thuth@redhat.com>
Signed-off-by: Christian Borntraeger <borntraeger@de.ibm.com>
(cherry picked from commit ee1563f42856f756c72bb0924109432d566c9efa)
Signed-off-by: Phillip Lougher <plougher@redhat.com>