[tools] kvm: selftests: Enable dirty_log_test on s390x
Message-id: <20190910154757.15199-19-thuth@redhat.com>
Patchwork-id: 271207
O-Subject: [RHEL8.2 virt BZ 1730412 PATCH 18/22] KVM: selftests: Enable dirty_log_test on s390x
Bugzilla: 1730412
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: David Hildenbrand <david@redhat.com>
RH-Acked-by: Peter Xu <peterx@redhat.com>

To run the dirty_log_test on s390x, we have to make sure that we
access the dirty log bitmap with little endian byte ordering and
we have to properly align the memslot of the guest.
Also all dirty bits of a segment are set once on s390x when one
of the pages of a segment are written to for the first time, so
we have to make sure that we touch all pages during the first
iteration to keep the test in sync here.
DEFAULT_GUEST_TEST_MEM needs an adjustment, too. On some s390x
distributions, the ELF binary is linked to address 0x80000000,
so we have to avoid that our test region overlaps into this area.
0xc0000000 seems to be a good alternative that should work on x86
and aarch64, too.

Acked-by: Paolo Bonzini <pbonzini@redhat.com>
Reviewed-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>
Link: https://lore.kernel.org/r/20190731151525.17156-4-thuth@redhat.com
Signed-off-by: Christian Borntraeger <borntraeger@de.ibm.com>
(cherry picked from commit a049a377164c3eca7d5323cd9896b64d2120c415)
Signed-off-by: Phillip Lougher <plougher@redhat.com>