[md] md: allow last device to be forcibly removed from RAID1/RAID10.
Message-id: <20190912155901.19857-7-ncroxon@redhat.com>
Patchwork-id: 271254
O-Subject: [RHEL8 PATCH 06/15] md: allow last device to be forcibly removed from RAID1/RAID10.
Bugzilla: 1730670
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Heinz Mauelshagen <heinzm@redhat.com>
RH-Acked-by: Xiao Ni <xni@redhat.com>

From: Guoqing Jiang <jgq516@gmail.com>

When the 'last' device in a RAID1 or RAID10 reports an error,
we do not mark it as failed.  This would serve little purpose
as there is no risk of losing data beyond that which is obviously
lost (as there is with RAID5), and there could be other sectors
on the device which are readable, and only readable from this device.
This in general this maximises access to data.

However the current implementation also stops an admin from removing
the last device by direct action.  This is rarely useful, but in many
case is not harmful and can make automation easier by removing special
cases.

Also, if an attempt to write metadata fails the device must be marked
as faulty, else an infinite loop will result, attempting to update
the metadata on all non-faulty devices.

So add 'fail_last_dev' member to 'struct mddev', then we can bypasses
the 'last disk' checks for RAID1 and RAID10, and control the behavior
per array by change sysfs node.

Signed-off-by: NeilBrown <neilb@suse.de>
[add sysfs node for fail_last_dev by Guoqing]
Signed-off-by: Guoqing Jiang <guoqing.jiang@cloud.ionos.com>
Signed-off-by: Song Liu <songliubraving@fb.com>
(cherry picked from commit 9a567843f7ce0037bfd4d5fdc58a09d0a527b28b)
Signed-off-by: Phillip Lougher <plougher@redhat.com>