[sound] ALSA: ASoC: hdmi-codec: remove reference to the current substream
Message-id: <15693081207948942-274-patchset-jkysela@redhat.com>
Patchwork-id: 273431
O-Subject: [RHEL 8.2 PATCH 0273/1298] ALSA: ASoC: hdmi-codec: remove reference to the current substream
Bugzilla: 1738610
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

If the hdmi-codec is on a codec-to-codec link, the substream pointer
it receives is completely made up by snd_soc_dai_link_event().
The pointer will be different between startup() and shutdown().

The hdmi-codec complains when this happens even if it is not really a
problem. The current_substream pointer is not used for anything useful
apart from getting the exclusive ownership of the device.

Remove current_substream pointer and replace the exclusive locking
mechanism with a simple variable and some atomic operations.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 3fcf94ef4d418668fa66e33ce9aabb05689b55f6)
Bugzilla: 1738610
Signed-off-by: Phillip Lougher <plougher@redhat.com>