[sound] ALSA: ASoC: hdmi-codec: remove reference to the dai drivers in the private data
Message-id: <15693081207948942-275-patchset-jkysela@redhat.com>
Patchwork-id: 273395
O-Subject: [RHEL 8.2 PATCH 0274/1298] ALSA: ASoC: hdmi-codec: remove reference to the dai drivers in the private data
Bugzilla: 1738610
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Keeping the a pointer to the dai drivers is not necessary. It is not used
by the hdmi_codec after the probe.

Even if it was used, the 'struct snd_soc_dai_driver' can accessed through
the 'struct snd_soc_dai' so keeping the pointer in the private data
structure is not useful.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 1de005d47d90343666c5cc50a50929e05e52baac)
Bugzilla: 1738610
Signed-off-by: Phillip Lougher <plougher@redhat.com>