[sound] ALSA: ASoC: rt5677-spi: Handle over reading when flipping bytes
Message-id: <15693081207948942-315-patchset-jkysela@redhat.com>
Patchwork-id: 273440
O-Subject: [RHEL 8.2 PATCH 0314/1298] ALSA: ASoC: rt5677-spi: Handle over reading when flipping bytes
Bugzilla: 1738610
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

There is a case when a we want to read a large number of bytes that
require a burst but is not a multiple of the word size (8). When this
happens rt5677_spi_reverse will run off the end of the buffer. The
solution is to tell spi_reverse the actual size of the destination and
stop if we reach it even if we have data left that we read.

Cc: Ben Zhang <benzh@chromium.org>
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 7b8164c1a29ce8ef91672c50ceac5c14475f5601)
Bugzilla: 1738610
Signed-off-by: Phillip Lougher <plougher@redhat.com>