[sound] ALSA: ASoC: rt5677: handle concurrent interrupts
Message-id: <15693081207948942-447-patchset-jkysela@redhat.com>
Patchwork-id: 273564
O-Subject: [RHEL 8.2 PATCH 0446/1298] ALSA: ASoC: rt5677: handle concurrent interrupts
Bugzilla: 1738610
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

The rt5677 driver writes to the IRQ control register within the IRQ
handler in order to flip the polarity of the interrupts that have been
signalled.  If an interrupt fires in the interval between the
regmap_read and the regmap_write, it will not trigger a new call to
rt5677_irq.

Add a bounded loop to rt5677_irq that keeps checking interrupts until
none are seen, so that any interrupts that are signalled in that
interval are correctly handled.

Signed-off-by: Ben Zhang <benzh@chromium.org>
Signed-off-by: Fletcher Woodruff <fletcherw@chromium.org>
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit df9091e9d3f4500bc6fb15f5d2a1c2614f67004c)
Bugzilla: 1738610
Signed-off-by: Phillip Lougher <plougher@redhat.com>