[sound] ALSA: ASoC: ti: davinci-mcasp: Correct slot_width posed constraint
Message-id: <15693081207948942-569-patchset-jkysela@redhat.com>
Patchwork-id: 273671
O-Subject: [RHEL 8.2 PATCH 0568/1298] ALSA: ASoC: ti: davinci-mcasp: Correct slot_width posed constraint
Bugzilla: 1738610
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

The slot_width is a property for the bus while the constraint for
SNDRV_PCM_HW_PARAM_SAMPLE_BITS is for the in memory format.

Applying slot_width constraint to sample_bits works most of the time, but
it will blacklist valid formats in some cases.

With slot_width 24 we can support S24_3LE and S24_LE formats as they both
look the same on the bus, but a a 24 constraint on sample_bits would not
allow S24_LE as it is stored in 32bits in memory.

Implement a simple hw_rule function to allow all formats which require less
or equal number of bits on the bus as slot_width (if configured).

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Link: https://lore.kernel.org/r/20190726064244.3762-2-peter.ujfalusi@ti.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 1e112c35e3c96db7c8ca6ddaa96574f00c06e7db)
Bugzilla: 1738610
Signed-off-by: Phillip Lougher <plougher@redhat.com>