[sound] ALSA: ASoC: ti: davinci-mcasp: Support for correct symmetric sample bits
Message-id: <15693081207948942-570-patchset-jkysela@redhat.com>
Patchwork-id: 273898
O-Subject: [RHEL 8.2 PATCH 0569/1298] ALSA: ASoC: ti: davinci-mcasp: Support for correct symmetric sample bits
Bugzilla: 1738610
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Implement custom snd_pcm_hw_rule to filter the available formats for the
second stream to make it symmetric and allow only formats which require
the same amount of bits on the bus as the running stream.

A simple constraint is not working correctly because for example:
the first stream is started with S24_LE

If we place 24 as constraint for the SAMPLE_BITS then the second stream
can not use S24_LE as it is physically 32bits.

If we would place 32 as constraint (physical width) then S32_LE would have
been allowed, but S24_3LE is not.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Link: https://lore.kernel.org/r/20190726064244.3762-3-peter.ujfalusi@ti.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 2448c813b12cf1fb08da1d20bdb2ae583cda2d6f)
Bugzilla: 1738610
Signed-off-by: Phillip Lougher <plougher@redhat.com>