[sound] ALSA: ASoC: create pcm for codec2codec links as well
Message-id: <15693081207948942-618-patchset-jkysela@redhat.com>
Patchwork-id: 273830
O-Subject: [RHEL 8.2 PATCH 0617/1298] ALSA: ASoC: create pcm for codec2codec links as well
Bugzilla: 1738610
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

At the moment, codec to codec links uses an ephemeral variable for
the struct snd_pcm_substream. Also the struct snd_soc_pcm_runtime
does not have real struct snd_pcm.

This might a problem if the functions used by a codec on codec to
codec link expect these structures to exist, and keep on existing
during the life of the codec.

For example, it is the case of the hdmi-codec, which uses
snd_pcm_add_chmap_ctls(). For the controls to works, the pcm and
substream must to exist.

This change is first step, it create pcm (and substreams) for codec
to codec links, in the same way as dpcm backend links.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Link: https://lore.kernel.org/r/20190725165949.29699-5-jbrunet@baylibre.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit a342031cdd0818cb0fbcb44798211c7a02c7ca27)
Bugzilla: 1738610
Signed-off-by: Phillip Lougher <plougher@redhat.com>