[sound] ALSA: ASoC: dmaengine: Make the pcm->name equal to pcm->id if the name is not set
Message-id: <15693081207948942-853-patchset-jkysela@redhat.com>
Patchwork-id: 273913
O-Subject: [RHEL 8.2 PATCH 0852/1298] ALSA: ASoC: dmaengine: Make the pcm->name equal to pcm->id if the name is not set
Bugzilla: 1738610
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Some tools use the snd_pcm_info_get_name() to try to identify PCMs or for
other purposes.

Currently it is left empty with the dmaengine-pcm, in this case copy the
pcm->id string as pcm->name.

For example IGT is using this to find the HDMI PCM for testing audio on it.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Reported-by: Arthur She <arthur.she@linaro.org>
Link: https://lore.kernel.org/r/20190906055524.7393-1-peter.ujfalusi@ti.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 2ec42f3147e1610716f184b02e65d7f493eed925)
Bugzilla: 1738610
Signed-off-by: Phillip Lougher <plougher@redhat.com>