[sound] ALSA: cs4281: remove redundant assignment to variable val and remove a goto
Message-id: <15693081207948942-905-patchset-jkysela@redhat.com>
Patchwork-id: 273889
O-Subject: [RHEL 8.2 PATCH 0904/1298] ALSA: cs4281: remove redundant assignment to variable val and remove a goto
Bugzilla: 1738610
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

The variable val is being assigned with a value that is never
read and it is being updated later with a new value. The
assignment is redundant and can be removed.  Also remove a
goto statement and a label and replace with a break statement.

Addresses-Coverity: ("Unused value")
Signed-off-by: Colin Ian King <colin.king@canonical.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 388b00f045180179fedff4b108d40a0192684cda)
Bugzilla: 1738610
Signed-off-by: Phillip Lougher <plougher@redhat.com>