[sound] ALSA: line6: Fix write on zero-sized buffer
Message-id: <15693081207948942-1256-patchset-jkysela@redhat.com>
Patchwork-id: 274297
O-Subject: [RHEL 8.2 PATCH 1255/1298] ALSA: line6: Fix write on zero-sized buffer
Bugzilla: 1738610
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

LINE6 drivers allocate the buffers based on the value returned from
usb_maxpacket() calls.  The manipulated device may return zero for
this, and this results in the kmalloc() with zero size (and it may
succeed) while the other part of the driver code writes the packet
data with the fixed size -- which eventually overwrites.

This patch adds a simple sanity check for the invalid buffer size for
avoiding that problem.

Reported-by: syzbot+219f00fb49874dcaea17@syzkaller.appspotmail.com
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 3450121997ce872eb7f1248417225827ea249710)
Bugzilla: 1738610
Signed-off-by: Phillip Lougher <plougher@redhat.com>