[sound] ALSA: usb-audio: Skip bSynchAddress endpoint check if it is invalid
Message-id: <15693081207948942-1267-patchset-jkysela@redhat.com>
Patchwork-id: 274389
O-Subject: [RHEL 8.2 PATCH 1266/1298] ALSA: usb-audio: Skip bSynchAddress endpoint check if it is invalid
Bugzilla: 1738610
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

The Linux kernel assumes that get_endpoint(alts,0) and
get_endpoint(alts,1) are eachothers feedback endpoints.
To reassure that validity it will test bsynchaddress to comply with that
assumption. But if the bsyncaddress is 0 (invalid), it will flag that as
a wrong assumption and return an error.
Fix: Skip the test if bSynchAddress is 0.
Note: those with a valid bSynchAddress should have a code quirck added.

Signed-off-by: Ard van Breemen <ard@kwaak.net>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 1b34121d9f26d272b0b2334209af6b6fc82d4bf1)
Bugzilla: 1738610
Signed-off-by: Phillip Lougher <plougher@redhat.com>