[sound] ALSA: usb-audio: Clean up check_input_term()
Message-id: <15693081207948942-1280-patchset-jkysela@redhat.com>
Patchwork-id: 274345
O-Subject: [RHEL 8.2 PATCH 1279/1298] ALSA: usb-audio: Clean up check_input_term()
Bugzilla: 1738610
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

The primary changes in this patch are cleanups of __check_input_term()
and move to a non-nested switch-case block by evaluating the pair of
UAC version and the unit type, as we've done for parse_audio_unit().
Also each parser is split into the function for readability.

Now, a slight behavior change by this cleanup is the handling of
processing and extension units.  Formerly we've dealt with them
differently between UAC1/2 and UAC3; the latter returns an error if no
input sources are available, while the former continues to parse.

In this patch, unify the behavior in all cases: when input sources are
available, it parses recursively, then override the type and the id,
as well as channel information if not provided yet.

Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit e0ccdef92653f8867e2d1667facfd3c23699f540)
Bugzilla: 1738610
Signed-off-by: Phillip Lougher <plougher@redhat.com>